+++ src/textReader.h
@@ -2,10 +2,10 @@
 #include <SdFat.h>
 
 extern FsFile file;
-extern unsigned long filePos;
-extern unsigned long nextFilePos;
+extern unsigned long long filePos;
+extern unsigned long long nextFilePos;
 
-extern unsigned long fileSize;
+extern unsigned long long fileSize;
 extern unsigned long bytesOnScreen;
 
 extern const byte maxCharPerLine;
@@ -13,9 +13,9 @@ extern const byte maxLines;
 
 void drawScrollbar(unsigned int startX, unsigned int startY, 
                    unsigned int width, unsigned int height,
-                   unsigned long startFill, unsigned long endFill, unsigned long maximum);
+                   unsigned long long startFill, unsigned long long endFill, unsigned long long maximum);
 
-void drawTextReaderScrollbar(unsigned long startFilePos, unsigned long endFilePos, unsigned long fileSize);
+void drawTextReaderScrollbar(unsigned long long startFilePos, unsigned long long endFilePos, unsigned long long fileSize);
 
 void drawDialog(const char* lines[], byte lineCount);
 
@@ -29,8 +29,8 @@ void noTextFilesScreen();
 
 void loadingTextFile(const char* filePath);
 
-unsigned long printFromLocation(unsigned int rows);
+unsigned long long printFromLocation(unsigned int rows);
 
-unsigned long getPosFromPage(unsigned int page, unsigned int rows);
+unsigned long long getPosFromPage(unsigned int page, unsigned int rows);
 
 unsigned int getMaxPage(unsigned int rows, unsigned long& crc);
